+++ lib/CHANGELOG.md
@@ -11,8 +11,11 @@ Added:
 - Parse more `VimNode` types: `Command` and `Flag`
 
 Changed:
-- Forked parse_module into parse_module_file and parse_module_str
+- Fork parse_module into parse_module_file and parse_module_str
 - Change parse_module_* return type `Vec<VimNode>`->`VimModule`
+
+Fixed:
+- Fix skipping after/ paths and some other subdirs like compiler/
 
 ## [0.2.x] - 2024-09-01
 Adds parse_plugin_dir